# docs/api/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "网格社区治理平台 API",
    "description": "网格社区治理平台后端接口文档 v1.2.0 — 完整覆盖所有控制器",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/parking/statistics": {
      "get": {
        "tags": [
          "停车管理"
        ],
        "summary": "车位统计",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/parking/spaces": {
      "get": {
        "tags": [
          "停车管理"
        ],
        "summary": "车位列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "停车管理"
        ],
        "summary": "添加车位",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/parking/spaces/{id}/status": {
      "put": {
        "tags": [
          "停车管理"
        ],
        "summary": "更新车位状态",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/parking/violations": {
      "get": {
        "tags": [
          "停车管理"
        ],
        "summary": "违停记录",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "停车管理"
        ],
        "summary": "添加违停记录",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/parking/violation-stats": {
      "get": {
        "tags": [
          "停车管理"
        ],
        "summary": "违停统计",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/parking/violations/{id}/dispatch": {
      "post": {
        "tags": [
          "停车管理"
        ],
        "summary": "派单处理违停",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/parking/violations/{id}/status": {
      "put": {
        "tags": [
          "停车管理"
        ],
        "summary": "更新违停状态",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/party/overview": {
      "get": {
        "tags": [
          "智慧党建"
        ],
        "summary": "党建总览统计",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/party/households": {
      "get": {
        "tags": [
          "智慧党建"
        ],
        "summary": "党员联户列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "智慧党建"
        ],
        "summary": "添加联户关系",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/party/households/{id}/visit": {
      "post": {
        "tags": [
          "智慧党建"
        ],
        "summary": "记录走访",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/party/activities": {
      "get": {
        "tags": [
          "智慧党建"
        ],
        "summary": "志愿服务活动列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "智慧党建"
        ],
        "summary": "添加志愿活动",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/party/activities/{id}/signup": {
      "post": {
        "tags": [
          "智慧党建"
        ],
        "summary": "报名志愿活动",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/party/meetings": {
      "get": {
        "tags": [
          "智慧党建"
        ],
        "summary": "三会一课列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "智慧党建"
        ],
        "summary": "添加会议记录",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/party/assessments": {
      "get": {
        "tags": [
          "智慧党建"
        ],
        "summary": "党员量化考核",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/party/assessments/generate": {
      "post": {
        "tags": [
          "智慧党建"
        ],
        "summary": "生成月度考核",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}